refactor: [hw9 & hw10] update proj structure
