+++ apps/weather-api/src/subscription/infrastructure/subscription.module.ts
@@ -4,8 +4,8 @@ import { SubscriptionRepository } from './subscription.repository';
 import { AbstractSubscriptionRepository } from '../domain-services/interfaces/subscription.interface';
 import { SubscriptionController } from '../application-services/subscription.controller';
 import { SubscriptionService } from '../domain-services/subscription.service';
-import { SubscriptionFacade } from '../public/subscription.facade';
-import { WeatherUpdateFacade } from '../public/weather-update.facade';
+import { SubscriptionFacade } from '../facade/subscription.facade';
+import { WeatherUpdateFacade } from '../facade/weather-update.facade';
 import { WeatherModule } from '@weather/infrastructure/weather.module';
 import { NotificationService } from '../domain-services/notification.service';
 import { EventBusModule } from '@shared/event-bus/infrastructure/event-bus.module';
